+++ src/hashgraph/badger_store.go
@@ -1,3 +1,5 @@
+// +build !mobile
+
 package hashgraph
 
 import (